requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.5
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
